pm-rapi/requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
boltons==25.0.0
celery==5.3.4
celery_longterm_scheduler==1.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
django-recurrence==1.11.1
django-username-email==2.5.7
djangorestframework==3.16.1
face==24.0.0
gevent==24.11.1
glom==24.11.0
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pendulum==3.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==6.4.0
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
zope.event==5.1.1
zope.interface==7.2
